Refactor resolvers with DDD(Domain-Driven Design)
