--- src/types/resolvers/Subscription/index.ts
@@ -1,7 +1,0 @@
-import { userSignedIn, userUpdated } from './User';
-export { USER_SIGNED_IN, USER_UPDATED } from './User';
-
-export const Subscription = [
-  userSignedIn,
-  userUpdated,
-];
